Bump memory limit for Prometheus containers to 1G (#73)
Towards https://github.com/giantswarm/giantswarm/issues/12251.

Memory requirements for Prometheus container depend on a couple of
factors which contribute to the fact that 200M was too low for some
installations causing Prometheus to get OOM killed as it hit the limit
shortly after startup. This in turn caused crash looping, which in turn
increased storage requirements for the WAL files.