Prevent deletion failures when source cert is already gone
When the generic resource tries to retrieve the desired state of an
object its reconciling it calls the `toCR` method. In case of
certificates (`Secret` resource) this method normally copies the data
from a source `Secret` (`default/$CLUSTER_ID-prometheus`). But when
deleting, this source may already be gone because we don't have a
finalizer on it. This would cause the `toCR` method to fail which in
turn caused the entire `EnsureDeleted` method to fail therefore leaving
the finalizer in place, leading to a zombie Cluster object that would
be continuously tried to be deleted but never succeeding.

Prevent this situation by avoiding failure in `toCR` method of
certificates resource when the source doesn't exist, but returning a
`Secret` object with an empty `Data` field.