Only use ObjectMeta on delete
Currently the generic resource deletion handler retrieves the desired
object only to read its name and namespace. For some resources this
causes a read of a dependant object from the cluster, which can lead
to problems when that is already gone, which in turn can prevent the
deletion from completing and leave zombie resources on the cluster.

This changes the generic interface to expect and use a `GetObjectMeta`
method which only returns the desired object metadata which doesn't
read from the cluster but provides information necessary for deletion.