Replace `bundleVersion` with `version`
Not sure how it got in here as it was replaced in all operators as part
of VOO refactoring.

Currently it's causing _prometheus-meta-operator_ to report incorrect
version of itself via `prometheus-meta-operator version` and via the
`giantswarm_build_info` metric (reports `version="0.0.1"`).

This fixes it by removing all references to `bundleVersion` with
`version` which is maintained in `project.go` via CI.