Don't copy TLS certs to per-cluster namespaces
This is unnecessary as it's enough for a single `Ingress` resource to
specify TLS configuration for a host, and in specific circumstances it
can in fact cause problems like an orphaned `Ingress` resource can
prevent an ingress-controller from picking up updated certificates.

This change removes copying of TLS certificates and duplicating TLS
configuration from `Ingress` resources, instead relying on the global
Prometheus `Ingress` (currently maintained in _g8s-prometheus_) being
the source-of-truth for the TLS configuration.

Closes: https://github.com/giantswarm/giantswarm/issues/14951