Fix missing Prom check when cluster is being deleted
This alert is based on `cluster_operator_cluster_status` metric which
may still show up after a cluster is marked for deletion with a label
`status="Deleting"`. This would cause a page when PMO has already
cleaned up the Prometheus shard for this cluster.

This patch excludes entries with `status="Deleting"` which should help
avoid paging in such cases.