Merge AlertManager global HTTP config with generated receivers
AlertManager doesn't do merging itself, which means that if we specify
HTTP config in the receivers (e.g. to specify auth credentials), then
the global config is not used for those receivers.

We want to preserve any options set in the global config which may
specify things like a proxy server to use, which affects the entire
environment.