Fix memory limit to avoid going over node capacity
VPA takes the maxAllowed to be based on memory request, that means the
request can be set to almost max memory available to the node, but since
the limit is calculated based on the request and multiplied by 1.2 it
means the limit can go higher than available memory.

Prevent this by lowering the maxAllowed value by the request-limit
coefficient.