Add `PrometheusHighMemoryUsage` alert
To page us during working hours if Prometheus container memory usage
(specifically the working set size) grows beyond 95% of total memory on
the node it's running on. This is dangerously close to OOM crash-loop
territory and likely needs to be addressed by either cleaning up some
data or upgrading to nodes with more memory.