Request team Atlas for review of release PRs
We want to be automatically requested for review of **release** PRs so
that we can advise in case there are reasons to postpone a release, e.g.
in case we want to bundle changes that are about to land, or we're aware
of a problem in the master branch, etc.